Create a deployment intended for Google Compute Engine
The deployment uses ESP (Extensible Service Proxy) in a reverse-proxy
configuration, with the JSON and WebSockets ports exposed via
`<HOSTNAME>:8080/jsrpc` and `<HOSTNAME>:8080/wsrpc` paths (resp). Also,
SSL is optionally used if connecting via port 8180 instead.

There is an attempt to expose the Parity Wallet (UI) and the Ethstats
dashboard as well, but it is currently not working.

 - Add configuration for nginx-based ESP as reverse-proxy
   NOTE: currently some config points are hardcoded for the specific
   Google Cloud project I'm using for development
 - Remove exposed ports in client & dashboard docker configs
 - Force-enable CLIENT and ETHSTATS options in parity-deploy.sh